+++ craftai/__init__.py
@@ -11,6 +11,7 @@ from . import errors
 from .client import CraftAIClient as Client
 from .interpreter import Interpreter
 from .time import Time
+from .formatters import format_property, format_decision_rule
 
 # Defining what will be imported when doing `from craftai import *`